@@ -1,12 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
+    <TargetFramework>net9.0-windows10.0.19041.0</TargetFramework>
     <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
     <RootNamespace>MediaPipe.GestureClassification</RootNamespace>
     <ApplicationManifest>app.manifest</ApplicationManifest>
     <Platforms>x86;x64;arm64</Platforms>
-    <RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
+    <RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
     <PublishProfile>win10-$(Platform).pubxml</PublishProfile>
     <UseWinUI>true</UseWinUI>
     <EnableMsixTooling>true</EnableMsixTooling>
@@ -116,13 +116,13 @@
 
   <ItemGroup>
     <PackageReference Include="CommunityToolkit.WinUI.UI.Controls" Version="7.1.2" />
-    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.3.230502000" />
-    <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.756" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.7.250606001" />
+    <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.26100.4654" />
 	  <PackageReference Include="Mediapipe.Net.Runtime.CPU" Version="1.0.0-alpha2" />
-	  <PackageReference Include="OpenCvSharp4" Version="4.7.0.20230115" />
-	  <PackageReference Include="OpenCvSharp4.Windows" Version="4.7.0.20230115" />
-	  <PackageReference Include="System.Drawing.Common" Version="7.0.0" />
-	  <PackageReference Include="OpenCvSharp4.Extensions" Version="4.7.0.20230115" />
+	  <PackageReference Include="OpenCvSharp4" Version="4.11.0.20250507" />
+	  <PackageReference Include="OpenCvSharp4.Windows" Version="4.11.0.20250507" />
+	  <PackageReference Include="System.Drawing.Common" Version="9.0.7" />
+	  <PackageReference Include="OpenCvSharp4.Extensions" Version="4.11.0.20250507" />
     <Manifest Include="$(ApplicationManifest)" />
   </ItemGroup>
 
